creates testdata package and extracts templates
